Move utils into separate folder
